BACKEND/requirements.txt
fastapi
pytest
uvicorn
sqlalchemy
psycopg2-binary
anthropic
python-dotenv
pre-commit
unidecode
requests
youtube_transcript_api
pytube
PyPDF2
clerk
python-multipart